ditched the overly complicated ImageGradient component in favor of a more direct solve. Series and Studies now have header gradients over their images 💥
